refactor: 입력과 결과 화면 패키지 변경
